Regression: Fix users raw model (#17204)
